WIP Microsserviço Pagamento
Endpoint de transferencia e pagamento de boletos pronto

+++ Pagamento/models/payment.py
@@ -1,7 +1,6 @@
 import uuid
 from pydantic import BaseModel
-
 class Payment(BaseModel):
-    value: int
+    value: float
     sender: int
     receiver: int
